+++ gatsby-config.js
@@ -8,6 +8,7 @@ const path = require('path');
 module.exports = {
   siteMetadata: {
     siteUrl: 'https://visiosto.netlify.app',
+    alternativeUrls: ['https://visiosto.fi', 'https://visiosto.com', 'https://visiosto.eu'],
     title: 'Visiosto',
     description:
       'Pilvilinnojen maalauksen aika on ohi, me rakennamme sen, mitä toivot. Taitamme sinun visiosi – linssimme läpi.',