+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 - Deploy status to the README.
 - Localized `404` error pages.
+- Redirects for different URLs.
 
 ### Changed